+++ packages/jest-validate/src/condition.ts
@@ -3,8 +3,6 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @flow
  */
 
 const toString = Object.prototype.toString;
@@ -23,6 +21,7 @@ function validationConditionSingle(option: any, validOption: any): boolean {
 export function getValues(validOption: any) {
   if (
     Array.isArray(validOption) &&
+    // @ts-ignore
     validOption[MULTIPLE_VALID_OPTIONS_SYMBOL]
   ) {
     return validOption;
@@ -34,9 +33,11 @@ export function validationCondition(option: any, validOption: any): boolean {
   return getValues(validOption).some(e => validationConditionSingle(option, e));
 }
 
+// TODO: This should infer the types of its arguments, and return a union type of the types
+// See https://github.com/Microsoft/TypeScript/issues/5453
 export function multipleValidOptions(...args: Array<any>) {
   const options = [...args];
-  // $FlowFixMe
+  // @ts-ignore
   options[MULTIPLE_VALID_OPTIONS_SYMBOL] = true;
   return options;
 }